requirements.txt
apturl==0.5.2
astroid==2.5.1
backcall==0.2.0
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
checkbox-support==0.22
cryptography==1.2.3
cycler==0.10.0
decorator==4.4.2
defer==1.0.6
easyocr==1.3.0.1
feedparser==5.1.3
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.0
imageio==2.9.0
imutils==0.5.4
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
Jinja2==2.8
joblib==1.0.1
kiwisolver==1.3.1
language-selector==0.1
lazy-object-proxy==1.6.0
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
matplotlib==3.3.4
mccabe==0.6.1
networkx==2.5
numpy==1.20.1
oauthlib==1.0.3
onboard==1.2.0
opencv-python==4.5.1.48
padme==1.1.1
pandas==1.2.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
plainbox==0.25
prompt-toolkit==3.0.17
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
Pygments==2.8.1
pygobject==3.20.0
PyJWT==1.3.0
pylint==2.7.2
pyparsing==2.0.3
PyQt5==5.9.2
pytesseract==0.3.7
python-apt==1.1.0b1+ubuntu0.16.4.11
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian==0.1.27
python-systemd==231
pytz==2021.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.4.1
reportlab==3.3.0
requests==2.9.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
selenium==3.141.0
sip==4.19.8
six==1.10.0
system-service==0.3
threadpoolctl==2.1.0
tifffile==2021.3.17
toml==0.10.2
torch==1.8.0+cpu
torchaudio==0.8.0
torchvision==0.9.0+cpu
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
wcwidth==0.2.5
wrapt==1.12.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
